States/Input 7-7-2022
Added functions working with states to store user input rather than have the static expenses shown.
Also added a dropdown to filter the year shown; the dropdown does not actually affect the webpage as of yet.
Removed unnecessary code i.e. the button that I used to check if the code was working as intended