7-8-2022
Rendered the list of expenses from a static hard-coded list to a user-inputted list.
Dropdown filter to show expenses per year is now functional.
Implemented a statement to inform users that there is no expenses found within a given year they have selected in the dropdown, if none exist.
Made Expenses.js leaner by creating ExpensesList.js
Added a button to open the form to launch the user input for new expenses with the tag of "Add New Expenses" and added a "Cancel" button to revert back and close the form.
Added a chart to serve as a visualization that allows users to see how much they spent per month.