7-8-2022 (3)
Changed the color scheme of the chart to match the rest of the website.
Edited the static items to allow the user to not deal with items that already exist.
Made the starting date 2022 to better reflect the present, also allowing users to immediately see results as we would generally be putting in expenses for the current year.